SI Test fixes
Summary:
dns was flaky and is resolved
resolving flaky vip tests with a retry
updating to shakedown supported wait for service calls
separating out the tests for root and mom. this is mainly due to ssh issues with python.

this fixes all SI test in the marathon repo... however there are SI tests that still fail:(

* ssh failure - is being investigated but I broke out the tests into root and mom tests to "hide" this failure
* neo4j in the root universe test fails and is a bug in shakedown

Test Plan: run tests

Reviewers: zen-dog, unterstein, jeschkies, meichstedt, jenkins

Reviewed By: zen-dog, jeschkies

Subscribers: marathon-dev

JIRA Issues: MARATHON_EE-1479

Differential Revision: https://phabricator.mesosphere.com/D747
